chore: Abstração no Board e Player
