+++ LagChessApplication/Extensions/PieceExtension.cs
@@ -1,5 +1,4 @@
-﻿using LagChessApplication.Domains;
-using LagChessApplication.Domains.Enums;
+﻿using LagChessApplication.Domains.Enums;
 using LagChessApplication.Interfaces;
 using System.Drawing;
 
@@ -25,7 +24,7 @@ namespace LagChessApplication.Extensions
                             new(-1, 0),
                     };
 
-                    return directions.Select(x => Point.Add(piece.Position, x)).Where(Board.IsInBoard).ToArray();
+                    return directions.Select(x => Point.Add(piece.Position, x)).Where(BoardExtension.IsInBoard).ToArray();
 
                 case PieceMoveStyleEnum.Straight:
                     var linearX = Enumerable.Range(1, 8).Where(x => piece.Position.X != x).Select(x => new Point(x, piece.Position.Y));
@@ -52,7 +51,7 @@ namespace LagChessApplication.Extensions
                         }
                     }
 
-                    return diagonal.Where(Board.IsInBoard).ToArray();
+                    return diagonal.Where(BoardExtension.IsInBoard).ToArray();
 
                 case PieceMoveStyleEnum.LShaped:
                     var knightMoves = new[]
@@ -67,7 +66,7 @@ namespace LagChessApplication.Extensions
                         new Point(piece.Position.X - 2, piece.Position.Y - 1)
                     };
 
-                    return knightMoves.Where(Board.IsInBoard).ToArray();
+                    return knightMoves.Where(BoardExtension.IsInBoard).ToArray();
 
                 case PieceMoveStyleEnum.OneAll:
 
@@ -89,7 +88,7 @@ namespace LagChessApplication.Extensions
                         }
                     }
 
-                    return oneStepMoves.Where(Board.IsInBoard).ToArray();
+                    return oneStepMoves.Where(BoardExtension.IsInBoard).ToArray();
 
                 case PieceMoveStyleEnum.All:
                     var linearMoves = GetPossibleMoves(piece, PieceMoveStyleEnum.Straight);
